refactor: components nomenclature
